requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
joblib>=1.3.0

# API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Dashboard
streamlit>=1.28.0

# Data & Networking
requests>=2.31.0
pyarrow>=13.0.0

# Scheduling
schedule>=1.2.0

# Config
pyyaml>=6.0.0

# Testing
pytest>=7.4.0
httpx>=0.24.0
